Refactor: score 컬럼 MemberDetail에 추가 및 기본값 0으로 설정완료

+++ src/main/java/com/starta/project/domain/quiz/service/CommentService.java
@@ -8,10 +8,11 @@ import com.starta.project.domain.quiz.entity.Comment;
 import com.starta.project.domain.quiz.entity.Quiz;
 import com.starta.project.domain.quiz.repository.CommentRepository;
 import com.starta.project.domain.quiz.repository.QuizRepository;
-import com.starta.project.global.messageDto.MsgDataResponse;
 import com.starta.project.global.messageDto.MsgResponse;
 import lombok.RequiredArgsConstructor;
+import org.springframework.http.ResponseEntity;
 import org.springframework.stereotype.Service;
+import org.springframework.transaction.annotation.Transactional;
 
 @Service
 @RequiredArgsConstructor
@@ -19,13 +20,10 @@ public class CommentService {
 
     private final QuizRepository quizRepository;
     private final CommentRepository commentRepository;
-    private final MemberRepository memberRepository;
 
-    public MsgResponse createComment(CreateCommentRequestDto createCommentRequestDto) {
+    public MsgResponse createComment(CreateCommentRequestDto createCommentRequestDto, Member member) {
         Quiz quiz = quizRepository.findById(createCommentRequestDto.getId()).orElseThrow( ()
         -> new NullPointerException("해당 퀴즈가 없습니다. "));
-        Member member = memberRepository.findById(1L).orElseThrow( () ->
-                new NullPointerException("해당 멤버는 없는 멤버 입니다. "));
         Comment comment = new Comment();
         comment.set(quiz,createCommentRequestDto,member);
         commentRepository.save(comment);
@@ -33,17 +31,25 @@ public class CommentService {
     }
 
 
-    public MsgResponse updateComment(Long id, UpdateCommentResponseDto updateCommentResponseDto) {
+    @Transactional
+    public ResponseEntity<MsgResponse> updateComment(Long id, UpdateCommentResponseDto updateCommentResponseDto, Member member) {
         Comment comment = findComment(id);
+
+        if(!member.getId().equals(comment.getMember().getId()) ) {
+            return ResponseEntity.badRequest().body( new MsgResponse( "댓글을 작성한 유저만 수정 가능합니다. "));
+        }
         comment.update(updateCommentResponseDto.getContent());
         commentRepository.save(comment);
-        return new MsgResponse("댓글 수정을 성공했습니다. ");
+        return ResponseEntity.ok().body(new MsgResponse("댓글 수정을 성공했습니다. "));
     }
 
-    public MsgResponse deleteComment(Long id) {
+    public ResponseEntity<MsgResponse> deleteComment(Long id, Member member) {
         Comment comment = findComment(id);
+        if(!member.getId().equals(comment.getMember().getId()) ) {
+            return ResponseEntity.badRequest().body( new MsgResponse( "댓글을 작성한 유저만 삭제 가능합니다. "));
+        }
         commentRepository.delete(comment);
-        return new MsgResponse("댓글 삭제를 성공했습니다. ");
+        return ResponseEntity.ok().body(new MsgResponse("댓글 삭제를 성공했습니다. "));
     }
 
     private Comment findComment (Long id) {
